Update Vyatta/Interface.pm to v1.10.5 -- Take 2
Fix previously botched version sync.

The change in v1.10.5 is actually two lines. Instead of copying
over `Interface.pm` from squashfs and adding `wg` entry to
`%net_prefix`, I fixed two lines in the repo.  This was prone to various
editor errors and here we are.  This time, done it the way I should've
done in the first place.